listening-comp/requirements.txt
azure-cognitiveservices-speech>=1.31.0
chromadb
streamlit
python-dotenv
ffmpeg-python>=0.2.0
youtube-transcript-api>=0.6.2
boto3>=1.34.0
requests>=2.31.0
uuid>=1.30.0
